refactor(cli): flatten the telegram command module
Move the Telegram command domain to the neutral command root and update its internal exports without changing public CLI or API behavior.
