+++ turbo/apps/cli/src/commands/telegram/__tests__/bot-list.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { http, HttpResponse } from "msw";
-import { server } from "../../../../mocks/server";
+import { server } from "../../../mocks/server";
 import { listCommand } from "../bot/list";
 import chalk from "chalk";
