+++ turbo/apps/cli/src/commands/telegram/bot/list.ts
@@ -1,7 +1,7 @@
 import { Command } from "commander";
 import chalk from "chalk";
-import { listTelegramBots } from "../../../../lib/api/domains/integrations-telegram";
-import { withErrorHandler } from "../../../../lib/command/with-error-handler";
+import { listTelegramBots } from "../../../lib/api/domains/integrations-telegram";
+import { withErrorHandler } from "../../../lib/command/with-error-handler";
 import type { TelegramBotListItem } from "@okouai/api-contracts/contracts/integrations";
 
 function usernameLabel(bot: TelegramBotListItem): string {
